pass `void*` arguments to entry function
Based on documentation for `makecontext` this _appears_ to be the correct implementation, but there may be some glibc-specific behavior greasing the wheels here. `makecontext` initially suggests only `int`-sized arguments can be passed to the given function, but quickly makes mentions glibc has some accomodations for x86 systems. Additionally, the type-checker complained about the type of the entry function; seeminly at odds with the documentation. In any case, this seems to work as expected.