adjust timer
Initial tests with the previous timer did not work as expected since sleeping a process did not result in accumulation of `PROCESS_CPU`-time. The goal here was to support multiple threads sleeping _concurrently_; the fix being to switch to a `MONOTONIC` timer. While this achieves the desired effect, I'm not 100% the overall impact on moving off of the previous timer, perhaps there were some additional reasons to make use of that original clock.